pyproject.toml
[tool.poetry]
name = "terminal-echoes"
version = "0.1.0"
description = ""
authors = ["Carlos Moreno <morenort1402@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
